modify log time format
I'm working on a tool to merge logfiles from a bunch of IVS switches run under
mininet. Using microseconds instead of milliseconds improves the quality of the
merged logfile by reducing the amount of reordering. For example, if switch A
sends a packet to switch B and gets a reply back within the same millisecond,
the log may show A getting the reply before B sends it.

Changing the month name to a number lets me merge without parsing the date. If
using the month name is some kind of standard and changing it would cause
issues then I'm fine with reverting this part.