Introduce the DEPENDMODULE_HEADERS variable.
Modules that require the public headers of another module in order to compile
but don't want to compile that module itself can now add those modules to
the DEPENDMODULE_HEADERS variable.

The current version is not entirely correct as it grabs the $(mod)/module/inc
directory directly when it should probably honor the $(mod)_INCLUDES variable
exported by the module's own makefiles. That setting would have to be
teased out of the makefile structure first.