tools/update-submodules: add option to not run git pull
This option is useful for more complicated usecases when you're not pulling
from origin/master. The user can do the pull himself and still get the nice
commit message generated by the script.

Also added support for specifying the submodule paths on the command line,
useful to selectively update a subset of submodules.