builder: remove extra slashes in paths
RPM has a [bug][1] that breaks debuginfo extraction when a filename in the
debug section has double slashes. There's some random chance involved which is
why we hadn't seen this problem before.

The main source of double slashes is `$(dir $(lastword $(MAKEFILE_LIST)))`,
which has a trailing slash. This patch removes the slashes added when
concatenating those paths with filenames.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=304121