Add new --foreach-module option.
The --foreach-module will run through all module entries in the loaded or generated database and run a command against the dictionary keys in each module.
This can be used to run maintenance tasks on all modules in a given tree. For example:

    > modtool.py --load-from . --foreach-module 'make -C %(dir)s autogen' will rerun autogen on under the current directory.

The command receives the module dict for variable replacement but the command can be anything.